fix: #286 admin path problem & chore: move config into folders
