+++ src/main/java/com/umc/linkyou/config/common/WebConfig.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.config;
+package com.umc.linkyou.config.common;
 
 import com.umc.linkyou.validation.annotation.ApiAdmin;
 import com.umc.linkyou.validation.annotation.ApiManager;
@@ -13,16 +13,19 @@ public class WebConfig implements WebMvcConfigurer {
 
     @Override
     public void configurePathMatch(PathMatchConfigurer configurer) {
-        // ApiV1 어노테이션이 있으면 /api/v1 추가
-        configurer.addPathPrefix("/api/v1",
-                c -> c.isAnnotationPresent(ApiV1.class));
-        configurer.addPathPrefix("/api/v2",
-                c -> c.isAnnotationPresent(ApiV2.class));
+        // ApiV1 어노테이션이 있으면 /api/v1/admin 추가
         configurer.addPathPrefix("/api/v1/admin",
                 c -> c.isAnnotationPresent(ApiAdmin.class));
-        // 매니저 전용 접두사 추가
+
         configurer.addPathPrefix("/api/v1/manage",
                 c -> c.isAnnotationPresent(ApiManager.class));
+
+        // 2. 그 다음 일반적인 경로 배치
+        configurer.addPathPrefix("/api/v1",
+                c -> c.isAnnotationPresent(ApiV1.class) && !c.isAnnotationPresent(ApiAdmin.class));
+
+        configurer.addPathPrefix("/api/v2",
+                c -> c.isAnnotationPresent(ApiV2.class));
     }
 }
